Added functions to actually process and render the code
I've expanded render_code() to actually return the correct HTML, which gets displayed on the page.

I've moved the list of lexers into config.py, as that's most likely to change on a regular basis.

I've moved the custom styles into an external CSS file, and created a file for the basic Pygments styles.