general tidy up and consolidation of code
All the Python has been moved into a single file `pygmentizr.py` since this is only a small script.

The CSS is now generated on the fly depending on the theme (which can be selected in `pygmentizr.py`.

Minor tweaks to the HTML to make it look slightly better on small screens.